requirements.txt
# AI Agent RAG Backend - Python Dependencies
# Production-ready version for Azure Deployment

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Azure OpenAI
openai>=1.12.0

# Azure AI Search
azure-search-documents==11.4.0
azure-identity>=1.15.0

# Data Processing
numpy>=1.24.0

# HTTP Client
httpx>=0.26.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Production Server
gunicorn==21.2.0
